+++ src/Playthrough.ts
@@ -1,5 +1,6 @@
 import EventEmitter from "eventemitter3";
-import io from "socket.io-client";
+import * as Colyseus from "colyseus.js";
+import jwtDecode from "jwt-decode";
 
 import * as api from "./api";
 
@@ -25,68 +26,87 @@ type PlaythroughEvents = {
 class Playthrough extends EventEmitter<PlaythroughEvents> {
   private token: string;
 
-  private socket: SocketIOClient.Socket | undefined;
+  private baseUrl?: string;
+
+  private client: Colyseus.Client | undefined;
+
+  private room: Colyseus.Room | undefined;
 
   private connectionStatus: ConnectionStatus = "disconnected";
 
   private activeConversations = new Map<ConversationId, Conversation>();
 
-  public constructor(token: string) {
+  public constructor(token: string, baseUrl?: string) {
     super();
     this.token = token;
+    this.baseUrl = baseUrl;
   }
 
   public createConversation(): ReturnType<typeof api.createConversation> {
-    return api.createConversation(this.token);
+    return api.createConversation(this.token, { baseUrl: this.baseUrl });
   }
 
   public createCharacterConversation(
     characterId: number,
   ): ReturnType<typeof api.createCharacterConversation> {
-    return api.createCharacterConversation(this.token, characterId);
+    return api.createCharacterConversation(this.token, characterId, {
+      baseUrl: this.baseUrl,
+    });
   }
 
   public getMessageHistory(
     conversationId?: string | undefined,
     minEventId?: string | undefined,
   ): ReturnType<typeof api.getMessageHistory> {
-    return api.getMessageHistory(this.token, conversationId, minEventId);
+    return api.getMessageHistory(this.token, conversationId, minEventId, {
+      baseUrl: this.baseUrl,
+    });
   }
 
   public getPlaythroughInfo(): ReturnType<typeof api.getPlaythroughInfo> {
-    return api.getPlaythroughInfo(this.token);
+    return api.getPlaythroughInfo(this.token, { baseUrl: this.baseUrl });
   }
 
   public setMemory(
     memoryIdOrRecallValue: number | string,
     saveValue: string | null,
   ): ReturnType<typeof api.setMemory> {
-    return api.setMemory(this.token, memoryIdOrRecallValue, saveValue);
+    return api.setMemory(this.token, memoryIdOrRecallValue, saveValue, {
+      baseUrl: this.baseUrl,
+    });
   }
 
   public setMood(
     characterIdOrName: number | string,
     modifier: Mood,
   ): ReturnType<typeof api.setMood> {
-    return api.setMood(this.token, characterIdOrName, modifier);
+    return api.setMood(this.token, characterIdOrName, modifier, {
+      baseUrl: this.baseUrl,
+    });
   }
 
   public restartFromEpisodeId(
     episodeId: number,
   ): ReturnType<typeof api.restartFromEpisodeId> {
-    return api.restartFromEpisodeId(this.token, episodeId);
+    return api.restartFromEpisodeId(this.token, episodeId, {
+      baseUrl: this.baseUrl,
+    });
   }
 
   public restartFromEpisodeIndex(
     episodeIndex: number,
   ): ReturnType<typeof api.restartFromEpisodeIndex> {
-    return api.restartFromEpisodeIndex(this.token, episodeIndex);
+    return api.restartFromEpisodeIndex(this.token, episodeIndex, {
+      baseUrl: this.baseUrl,
+    });
   }
 
   public restartFromEventId(
     eventId: string,
   ): ReturnType<typeof api.restartFromEventId> {
-    return api.restartFromEventId(this.token, eventId);
+    return api.restartFromEventId(this.token, eventId, {
+      baseUrl: this.baseUrl,
+    });
   }
 
   public joinConversation = (
@@ -116,13 +136,10 @@ class Playthrough extends EventEmitter<PlaythroughEvents> {
     return this.activeConversations.get(conversationId);
   };
 
-  public addOutgoingEvent = (
-    eventName: string,
-    ...eventData: unknown[]
-  ): void => {
-    if (this.socket) {
+  public addOutgoingEvent = (eventName: string, eventData: unknown): void => {
+    if (this.room) {
       if (this.connectionStatus === "connected") {
-        this.socket.emit(eventName, ...eventData);
+        this.room.send(eventName, eventData);
       } else {
         console.warn(
           `Event \`${eventName}\` was not sent as the socket was not ready. Wait for the \`connection-status\` event to be called with \`connected\` before sending events.`,
@@ -135,35 +152,44 @@ class Playthrough extends EventEmitter<PlaythroughEvents> {
     }
   };
 
-  public connect = (): void => {
-    const charismaUrl = api.getBaseUrl();
-    this.socket = io(`${charismaUrl}/play`, {
-      query: { token: this.token },
-      transports: ["websocket"],
-      upgrade: false,
+  public connect = async (): Promise<void> => {
+    const baseUrl = this.baseUrl || api.getGlobalBaseUrl();
+
+    if (!this.client) {
+      this.client = new Colyseus.Client(baseUrl.replace(/^http/, "ws"));
+    }
+
+    // eslint-disable-next-line camelcase
+    const { playthrough_id } = jwtDecode<{ playthrough_id: number }>(
+      this.token,
+    );
+
+    this.room = await this.client.joinOrCreate("chat", {
+      playthroughId: playthrough_id,
+      token: this.token,
     });
 
-    // Fired upon a successful reconnection.
-    this.socket.on("reconnect", this.onReconnect);
-    // Fired upon an attempt to reconnect.
-    this.socket.on("reconnecting", this.onReconnecting);
-    // Fired upon a disconnection.
-    this.socket.on("disconnect", this.onDisconnect);
-    // Fired when an error occurs.
-    this.socket.on("error", this.onError);
+    // // Fired upon a successful reconnection.
+    // this.socket.on("reconnect", this.onReconnect);
+    // // Fired upon an attempt to reconnect.
+    // this.socket.on("reconnecting", this.onReconnecting);
+    // // Fired upon a disconnection.
+    // this.socket.on("disconnect", this.onDisconnect);
+    // // Fired when an error occurs.
+    // this.socket.on("error", this.onError);
 
-    this.socket.on("status", this.onStatus);
-    this.socket.on("problem", this.onProblem);
-    this.socket.on("start-typing", this.onStartTyping);
-    this.socket.on("stop-typing", this.onStopTyping);
-    this.socket.on("message", this.onMessage);
-    this.socket.on("episode-complete", this.onEpisodeComplete);
+    this.room.onMessage("status", this.onStatus);
+    this.room.onMessage("problem", this.onProblem);
+    this.room.onMessage("start-typing", this.onStartTyping);
+    this.room.onMessage("stop-typing", this.onStopTyping);
+    this.room.onMessage("message", this.onMessage);
+    this.room.onMessage("episode-complete", this.onEpisodeComplete);
   };
 
   public cleanup = (): void => {
-    if (this.socket) {
-      this.socket.close();
-      this.socket = undefined;
+    if (this.room) {
+      this.room.leave();
+      this.room = undefined;
     }
   };
 
@@ -174,32 +200,32 @@ class Playthrough extends EventEmitter<PlaythroughEvents> {
     }
   };
 
-  private onReconnect = (): void => {
-    this.activeConversations.forEach((conversation) => {
-      conversation.reconnect().catch((err) => {
-        console.error(
-          `Something went wrong reconnecting to conversation:`,
-          err,
-        );
-      });
-    });
-  };
+  // private onReconnect = (): void => {
+  //   this.activeConversations.forEach((conversation) => {
+  //     conversation.reconnect().catch((err) => {
+  //       console.error(
+  //         `Something went wrong reconnecting to conversation:`,
+  //         err,
+  //       );
+  //     });
+  //   });
+  // };
 
-  private onReconnecting = (): void => {
-    this.changeConnectionStatus("connecting");
-  };
+  // private onReconnecting = (): void => {
+  //   this.changeConnectionStatus("connecting");
+  // };
 
-  private onDisconnect = (): void => {
-    this.changeConnectionStatus("disconnected");
-  };
+  // private onDisconnect = (): void => {
+  //   this.changeConnectionStatus("disconnected");
+  // };
 
   private onStatus = (): void => {
     this.changeConnectionStatus("connected");
   };
 
-  private onError = (error: unknown): void => {
-    this.emit("error", error);
-  };
+  // private onError = (error: unknown): void => {
+  //   this.emit("error", error);
+  // };
 
   private onProblem = (problem: { type: string; error: string }): void => {
     this.emit("problem", problem);